Expose VM#addSound publicly
